split out help panel
