+++ crates/client/src/ui/panels/mod.rs
@@ -1,0 +1,3 @@
+pub mod help;
+
+pub use help::render_help_panel;
